+++ packages/orq-rc/RELEASES.md
@@ -6619,3 +6619,13 @@ Based on:
 - [typescript v4.3.0-rc.48] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.48 - packages/orq-rc
+
+## 2026-02-10 14:37:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.3 (2.814.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.49] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.49 - packages/orq-rc